Web: fix document node searching (#46506)
* Delete unused nodes files

* Move search panel out of unified to teleport/src/component

The search panel in unified dir  wasn't that much different
from the one already existing in teleport/src/component
other than the latter one displayed a page indicator.

DocumentNodes and other places now use the same kind of server
side search panel.

* Get rid of input search auto focus

The auto focus can steal focus from other elements
