+++ web/packages/teleport/src/Console/DocumentNodes/useNodes.ts
@@ -18,7 +18,6 @@
 
 import { useEffect } from 'react';
 
-import { sortLogins } from 'teleport/Nodes/useNodes';
 import {
   useUrlFiltering,
   useServerSidePagination,
@@ -103,3 +102,13 @@ export default function useNodes({ clusterId, id }: stores.DocumentNodes) {
     ...paginationProps,
   };
 }
+
+// sort logins by making 'root' as the first in the list
+export const sortLogins = (logins: string[]) => {
+  const noRoot = logins.filter(l => l !== 'root');
+  noRoot.sort();
+  if (noRoot.length === logins.length) {
+    return logins;
+  }
+  return ['root', ...noRoot];
+};
