+++ web/packages/teleport/src/components/ServersideSearchPanel/SearchInput.tsx
